Refactor to use one single configProvider context
