+++ src/contexts/Guilds/config/hook.ts
@@ -1,0 +1,5 @@
+import React from 'react';
+import { GuildConfigContext, ConfigContext } from './context';
+
+export const useGuildConfig = () =>
+  React.useContext<ConfigContext>(GuildConfigContext);
