--- src/contexts/Guilds/guildConfig/index.ts
@@ -1,5 +1,0 @@
-export {
-  GuildConfigContext,
-  GuildConfigProvider,
-  useGuildConfigContext,
-} from './guildConfig';
