+++ src/contexts/Guilds/index.ts
@@ -1,5 +1,4 @@
 export * from './filters';
 export * from './proposals';
-export * from './guildConfig';
-export * from './providerService';
+export * from './config';
 export * from './rootContextProvider';
