--- src/contexts/Guilds/providerService/index.ts
@@ -1,5 +1,0 @@
-export {
-  GuildServiceProvider,
-  GuildServiceContext,
-  useGuildProviderService,
-} from './providerService';
