+++ src/contexts/Guilds/rootContextProvider.tsx
@@ -1,22 +1,16 @@
 import React from 'react';
 
-import {
-  FilterProvider,
-  ProposalsProvider,
-  GuildConfigProvider,
-  GuildServiceProvider,
-} from '.';
+import { FilterProvider, ProposalsProvider } from '.';
+import { GuildConfigProvider } from './config';
 
 export const GuildsContextProvider: React.FC<{
   children: React.ReactNode;
 }> = ({ children }) => {
   return (
-    <GuildServiceProvider>
-      <GuildConfigProvider>
-        <FilterProvider>
-          <ProposalsProvider>{children}</ProposalsProvider>
-        </FilterProvider>
-      </GuildConfigProvider>
-    </GuildServiceProvider>
+    <GuildConfigProvider>
+      <FilterProvider>
+        <ProposalsProvider>{children}</ProposalsProvider>
+      </FilterProvider>
+    </GuildConfigProvider>
   );
 };
